+++ controller/controller.go
@@ -36,6 +36,7 @@ import (
 	"github.com/giantswarm/operatorkit/controller/context/reconciliationcanceledcontext"
 	"github.com/giantswarm/operatorkit/controller/context/resourcecanceledcontext"
 	"github.com/giantswarm/operatorkit/controller/context/updateallowedcontext"
+	"github.com/giantswarm/operatorkit/controller/internal/recorder"
 	"github.com/giantswarm/operatorkit/controller/internal/sentry"
 	"github.com/giantswarm/operatorkit/resource"
 )
@@ -96,6 +97,7 @@ type Controller struct {
 	k8sClient            k8sclient.Interface
 	logger               micrologger.Logger
 	newRuntimeObjectFunc func() pkgruntime.Object
+	event                recorder.Interface
 	resources            []resource.Interface
 	selector             Selector
 
@@ -159,6 +161,17 @@ func New(config Config) (*Controller, error) {
 		}
 	}
 
+	var eventRecorder recorder.Interface
+	{
+		c := recorder.Config{
+			K8sClient: config.K8sClient,
+
+			Component: config.Name,
+		}
+
+		eventRecorder = recorder.New(c)
+	}
+
 	var sentryClient sentry.Interface
 	{
 		c := sentry.Config{
@@ -178,6 +191,7 @@ func New(config Config) (*Controller, error) {
 		logger:               config.Logger,
 		selector:             config.Selector,
 		newRuntimeObjectFunc: config.NewRuntimeObjectFunc,
+		event:                eventRecorder,
 		resources:            config.Resources,
 
 		backOffFactory:         func() backoff.Interface { return backoff.NewMaxRetries(7, 1*time.Second) },
@@ -241,8 +255,24 @@ func (c *Controller) Reconcile(req reconcile.Request) (reconcile.Result, error) 
 		ctx = setLoggerCtxValue(ctx, loggerKeyController, c.name)
 	}
 
-	res, err := c.reconcile(ctx, req)
+	obj := c.newRuntimeObjectFunc()
+	err := c.k8sClient.CtrlClient().Get(ctx, req.NamespacedName, obj)
+	if errors.IsNotFound(err) {
+		// At this point the controller-runtime cache dispatches a runtime object
+		// which is already being deleted, which is why it cannot be found here
+		// anymore. We then likely perceive the last delete event of that runtime
+		// object and it got purged from the controller-runtime cache. We do not
+		// need to log these errors and just stop processing here in a more graceful
+		// way.
+		return reconcile.Result{}, nil
+	} else if err != nil {
+		return reconcile.Result{}, microerror.Mask(err)
+	}
+
+	res, err := c.reconcile(ctx, req, obj)
 	if err != nil {
+		// Microerror creates an error event on the object when kind and description is set.
+		c.event.Emit(ctx, obj, err)
 		errorGauge.Inc()
 		c.sentry.Capture(ctx, err)
 		c.logger.LogCtx(ctx, "level", "error", "message", "failed to reconcile", "stack", microerror.JSON(err))
@@ -369,22 +399,8 @@ func (c *Controller) deleteFunc(ctx context.Context, obj interface{}) error {
 	return nil
 }
 
-func (c *Controller) reconcile(ctx context.Context, req reconcile.Request) (reconcile.Result, error) {
-	obj := c.newRuntimeObjectFunc()
-	err := c.k8sClient.CtrlClient().Get(ctx, req.NamespacedName, obj)
-	if errors.IsNotFound(err) {
-		// At this point the controller-runtime cache dispatches a runtime object
-		// which is already being deleted, which is why it cannot be found here
-		// anymore. We then likely perceive the last delete event of that runtime
-		// object and it got purged from the controller-runtime cache. We do not
-		// need to log these errors and just stop processing here in a more graceful
-		// way.
-		return reconcile.Result{}, nil
-	} else if err != nil {
-		return reconcile.Result{}, microerror.Mask(err)
-	}
-
-	ctx, err = c.initCtx(ctx, obj)
+func (c *Controller) reconcile(ctx context.Context, req reconcile.Request, obj interface{}) (reconcile.Result, error) {
+	ctx, err := c.initCtx(ctx, obj)
 	if err != nil {
 		return reconcile.Result{}, microerror.Mask(err)
 	}