+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## Added
+
+- Support writing Kubernetes error events when returning microerror with kind and description.
+
 ## [1.1.0] - 2020-06-22
 
 ## Added